Drop references to now-unspecified DPR HTTP header (#415)
This change drops all references in all MDN content to the DPR HTTP
header. The DPR HTTP header is now completely unspecified; that is, it
is no longer defined in any current specification.

So this change also marks the web/http/headers/dpr article with Obsolete
and Non-standard notices and tags.