Using two selectors description change (#1876)
* Using two selectors description

"__Using two selectors at the same time will not work. For example, :not(.foo,bar) is invalid. Instead, :not(.foo):not(.bar) should be used.__". 
This line in the description is no longer true. 

At least Chrome 88, Firefox 84, Edge 88 support this compound selector. [Code example](http://bit.ly/39qFXFQ)
Although, Opera 87 still doesn't support it. 

So we can either remove this line altogether or change it, so it won't say that this is an invalid way to use `:not` pseudo-class.

* Update files/en-us/web/css/_colon_not/index.html

Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>

Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>