Make links to CSS Grid &Flexbox a bit more aesthetic (#2287)
Make links to CSS Grid and Flexbox a bit more aesthetic and thus more readable. 
As for now, the wide link inside the list item which refers to the flexbox distracts the user from reading the whole sentence. 
The same goes for the list item with the grid link in it.