Remove specific line reference to script tag in HTML file (#4073)
* Refer index.js file to line 41 instead of line 40

The documentation currently refers to the index.js file at line 40 in file personal-greeting.html. However, it looks like the index.js file is actually being referenced on line 41.

* Remove specific line reference to index.js script

The original line reference to the index.js file in the HTML was incorrect. But because the line reference will likely continue changing over time, a specific line number reference was removed.

* Escape HTML for script tag

Co-authored-by: wbamberg <will@bootbonnet.ca>