Remove the note about "opinions" (#6708)
It's unclear what "opinions" are mentioned here. For my own part, I've not come across any differing opinions about this use of `role="presentation"` ... it works, it does what it's supposed to. If there are genuine concerns/issues with it, keep the note and include them. Otherwise, this only spread doubt unnecessarily.