Fixed grammar (#6739)
The sentence, "Every part of MDN (docs, demos, and the site itself) is created by an open community of developers." is an emendation of the original. The original had the subject "part" failing to agree with the verb "are".  This fails because the subject is singular and the verb was plural.