Remove rotationRate from WebGL 2D example (#7369)
The variable `rotationRate` is not used in the demo. This variable is obsolete and probably was substituted by  `degreesPerSecond`

I've found no other mention of this variable in the WebGL section of this repo: https://github.com/mdn/content/search?q=rotationRate
I also tested the code without the variable and it ran correctly.