Updates to the Box Model: margin collapsing (#6946)
* added clarity

Also, I've never seen a present participle or a gerund used to signify the name of a categorial concept before. It sounds odd to say "margin collapsing" as if it were such a concept. I may very well be wrong here, since my skills are with English and not with web development.

* Update files/en-us/learn/css/building_blocks/the_box_model/index.html

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Update files/en-us/learn/css/building_blocks/the_box_model/index.html

* Update files/en-us/learn/css/building_blocks/the_box_model/index.html

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>
Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>