docs: use simpler language for row- and column-reverse descriptions (#8762)
The word "permuted" was being used. This tripped me up so much while reading that I had to stop and open this PR.

This is not very approachable language. It's not a commonly used word, so let's try to improve the reading level for these descriptions. 

Open to suggestions or improvements for my proposed change! 🙂