Fix fibanacci example (#9625)
In the "Performing computations in the background" section there is a strange example about the use of Web Worker to calculate, in a background thread, a fibonacci number. This example doesn´t work. Also, there is a link to a web page of this example and web is working because it is another example. I replace the "strange" example by web page example.