Remove not-so-bad bad example of empty statement (#10114)
I stared at this description, and the stated reason it's bad. I don't see the problem
with this code -- specifically, it would have the same behavior if braces were used
instead of the empty-statement semicolon. (If not, if the empty-statement semicolon
in case three actually meant "end the if/else chain here and start a new statement",
then the subsequent "else if" wouldn't _parse_. So... that can't be it.)