Note a11y implications of pseudo-elements (#11348)
* Note a11y implications of pseudo-elements

Using `::after` pseudo-elements to add meaningful content to a page is a violation of accessibility guidelines, as it is not reliably accessible by screen readers. It is helpful to include such information in the basic documentation on the element. (If not this content, some discussion at least of the a11y implications would be helpful.)

https://accessibleweb.com/knowledgebase/perceivable/1-3-adaptable/01-03-01-info-and-relationships-level-a/
https://accessibleweb.com/question-answer/how-is-css-pseudo-content-treated-by-screen-readers/

* created a separate section for a11y concerns

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>