Confusing description of form property (#11418)
* Confusing description of form property

In the documentation, it is written:
"Is a HTMLFormElement reflecting the form that this button is associated with. If the button is a descendant of a form element, then this attribute is the ID of that form element."
So, which is it? an object or an id? After checking it turns out `button.form` always returns a `HTMLFormElement` object.

```
var form = document.createElement('form');
var button = document.createElement('button');
form.appendChild( button );
button.form instanceof( HTMLFormElement ); // true
```

* Apply suggestions from code review

Co-authored-by: Joe Medley <jmedley@google.com>

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>
Co-authored-by: Joe Medley <jmedley@google.com>