Update index.md (#11484)
Correct the description by including a restriction from the specification. This prevents users from inferring that it is possible to add a token to the list when the token is already a member of the list, perhaps causing the list to contain duplicate members. Making that inference could cause them to do the unnecessary work of checking the list for the existence of a member before adding it.