Make GeneratorFunction doc be clear it’s not a property of global object (#14807)
* Make GeneratorFunction doc be clear it’s not a property of global object

Fixes https://github.com/mdn/content/issues/14781

GeneratorFunction is an intrinsic object in the ES spec but it’s not a
exposed as a property of the global object. So there is no exposed
constructor with the name GeneratorFunction(), and we should make that
clear in the article. And because there’s no such exposed constructor,
we should not include a Syntax section for it.

* Apply suggestions from code review

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Update files/en-us/web/javascript/reference/global_objects/generatorfunction/index.md

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>