Fixing external link anchor to specification (#15055)
* Fixing external link anchor to specification

The link to the CSS2 specification links to the anchor `#q19.0` which does not exist. I have fixed the anchor to `#frequencies`.

* Update files/en-us/web/css/frequency/index.md

Would it be worth adding a note just that no CSS property is using this property at the moment? Someone stumbling across this page might wonder what the property is for.

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>