Remove outdated information about `xlink:href` (#15435)
* Remove outdated information about `xlink:href`

This warning suggests that `xlink:href` may still be necessary for cross-browser compatibility; however, the table at the bottom shows that the stable version of all browsers (including IE) now support `href`, so this warning is now misleading. In addition, it conflicts with the more accurate warning further down the page that `xlink:href` is deprecated.

* Restoring a portion of the warning

This may still be redundant with the warning below, but I restored it as it could be useful to anyone familiar with the old `xlink:href` pattern.