HTTP: Replace the last remaining manual spec tables (#16225)
A few articles in the HTTP tree with manual spec tables didn’t get
changed in the initial round of conversions. This change replaces those
last manual spec tables with {{Specifications}} macro calls.

Because some of the relevant specs aren’t in browser-specs, this change
adds them to the SpecData.json file. And in the case of the SourceMap
header, the spec probably *should* be in browser-specs, so see
https://github.com/w3c/browser-specs/issues/618 for that.

Related BCD PR: https://github.com/mdn/browser-compat-data/pull/16361