In the “Web app manifests” doc, replace manual spec table with inline link to spec (in the article intro) (#16250)
* In the “Web app manifests” doc, replace manual spec table

Note that in BCD, html.manifest is a directory, not a file. Given that,
there’s no way in BCD to associate a spec URL with the html.manifest
feature. And anyway for the “Web app manifests” MDN article — which is
an overview (non-reference) article — we don’t actually need a
Specifications table. So this change replaces the table with an inline
link to the spec (in the article intro).

* Drop Specifications section