Simplify is_overdue with bool() (#17206)
* Simplify is_overdue with bool()

Simplifies `is_overdue` using `bool()` for more "pythonic" code, and includes a note above it to explain the function to the reader. I included the note above as there's already one below, and I wanted to avoid there being two in a row.

* Remove unneeded conditional from is_overdue

Removes the if clause that I mistakenly left in. Went back to change the note, lost the original change, and then rewrote it incorrectly :)

* Apply suggestions from code review

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>