Fix a link in “Namespaces crash course” (#17591)
Fixes https://github.com/mdn/content/issues/17580

hasAttribute was (I think) first implemented browsers as part of DOM
Level 1 — even though it was not part of the DOM Level 1 spec. It didn’t
get specified formally until DOM Level 2.

Given all that, it’s not inaccurate to list it in the table in the
“Namespaces crash course” page as a DOM Level 1 element — while linking
to the DOM Level 2 spec for its definition.

So that’s what this change does.