Update index.md (#19391)
Summary:
1) Fixed 'og:image' path, which was using relative path, which isn't allowed in the open graph protocol:
https://indieweb.org/The-Open-Graph-protocol#Disallows_relative_URLs
2) Updated URL's from HTTP to HTTPS

Motivation:
I searched about what to add to the header for pwa and saw the 'og:image' meta tag which was missing in my HTML file. So I added it and found out that relative paths don't work. To prevent someone making eventually the same mistake, here a little fix.