Fix 'unexpected token' error in example JS code (#20725)
* Fix 'unexpected token' error in example JS code

This change relates to JS code in point 2 of the 'Making the action reusable' section. The JS object returned by the function 'selectOnFocus' has a semicolon in it that causes an 'unexpected token' error. Removing the semicolon fixes the problem and matches the code found on the REPL (https://svelte.dev/repl/d1fa84a5a4494366b179c87395940039?version=3.23.2).

* Update files/en-us/learn/tools_and_testing/client-side_javascript_frameworks/svelte_reactivity_lifecycle_accessibility/index.md

Co-authored-by: Joshua Chen <sidachen2003@gmail.com>