Update index.md (#20803)
I've changed the media query max-width from 12450 to 1250 because its easy to mislook the extra number.

When copying and pasting, and resizing the viewport, no changes are applied, and you enter a void of questions on why. 

Changing it to a more realistic value you encounter day to day will be easy to verify if the media query also works right away, since most if not all screens are 1250px in width for majority of devs. 

Making this chance will mean people won't scratch their head thinking why media queries aren't working, when in reality the docs had an extra number that devs overlooked.