Correct aeronautical misconceptions in 'Introduction to events' (#23072)
Pilots do not receive a 'signal' when they are cleared for take-off.
Pilots request take-off clearance over the radio and receive it over the
radio. Most importantly, no marshaller stands in front of the aircraft
on the runway when it is taking off, as the image seems to suggest!
If one really wants to use an aviation analogy here, I think a better
example of an event would be an engine failure, which would prompt the
pilot(s) to go through the engine failure checklist.