Precision on SSE listeners  (#23399)
* Precision on SSE listeners

It was not clear on which listener to use or not. the limitation usage is written after the sample. To understand which sample to use, it seems to me that making limitation before is more understandably. At first and second read, the two listeners seems to be equivalent where as they are exclusive.

* squash

* Update files/en-us/web/api/server-sent_events/using_server-sent_events/index.md

Co-authored-by: wbamberg <will@bootbonnet.ca>