Editorial updates to aria-controls content (#23960)
* Editorial updates to aria-controls content

Making some revisions to how the `aria-controls` attribute was described here.  The reason for this is due to some misunderstandings that developers had, and referenced this as their source.  

Specifically, this modifies instances of content that people had been interpreting as requirements, toning down the language a bit to leave the intent, but to more align with the spec.

swaps out the text example of a link with aria-controls, which is valid – but generally one would want developers to use a button in this situation.  Rather, though also not particularly common, the example was replaced with a mention of `role=scrollbar` - besides combobox, the only other role in ARIA that actually _requires_ the use of aria-controls.

Finally, the tab widget markup example was modified to use classes instead of the `hidden` attribute.  For reasons I've already alluded to with other updates I've made to ARIA tab-related articles, it'd require far more explanation to indicate the situations which `hidden` would not be considered an anti-pattern for this widget.  Rather than explain that here, which would be out of scope, replacing with classes that invoke the same idea that the inactive tabpanels are completley hidden, seems far less work.

* typo

* tick marks

* plural

---------

Co-authored-by: Estelle Weyl <estelle@weyl.org>