Myndex2023 1 (#25144)
* sRGB threshold

Updated the sRGB threshold to the correct value of 0.04045 which is also reflected in WCAG 2.1 as of May 2021. Also updated the links to the authoritative references.

* Links to luminance calculators

The Juicystudio calculator is very old, using incorrect terminology, and does not seem to be updated to the latest WCAG info. Replaced it with a link to Lea Verou's contrast tool which is more modern and easier to use. Lea Verou is an accredited color scientist.

* Add cone mosaic, and

This is some updates to the area relating to color perception, I tried very hard to keep the spirit of what and how it was written, simply trying to correct some errata. Also added an image of the cone mosaic to show the typical dispersement of cones, one for StandardVision and one for protanopia. Please let me know if you have any questions or if you feel it needs additional references or edits.

* Lint

* lint and a bit

* hmmm

Trading space keeps appearing weird

* Capitalization

OK I don't know if this link is gonna work, there's an offsite link to Wikipedia that uses camel case but the winter set it had to be lowercase only

Line 193   [wikicommons](https://commons.wikimedia.org/wiki/File:ConeMosaics.jpg)

* Delete ConeMosaics.jpg

* Revert "Delete ConeMosaics.jpg"

This reverts commit e9d3afb60a5971a4bc01e019555200bac421b36e.

* jpg

* Add files via upload

* Update files/en-us/web/accessibility/understanding_colors_and_luminance/index.md

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>

* Update files/en-us/web/accessibility/understanding_colors_and_luminance/index.md

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>

* Update files/en-us/web/accessibility/understanding_colors_and_luminance/index.md

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>

* Update files/en-us/web/accessibility/understanding_colors_and_luminance/index.md

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>

* prettier: changed quote styles

* extra line break (prettier)

* Apply suggestions from code review

---------

Co-authored-by: Estelle Weyl <estelle@weyl.org>
Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>