Update example usage of NodeList.length to actually use .length in example instead of for-of loop (#25470)
Update example usage of .length to actually use .length

Previously the example usage of .length as an iterator in a for loop was not accurately displayed. Instead of using .length as the iterator in the loop, a for-of loop was used instead. Though the for-of loop may have been better in this case, it does not display the use of .length. This change fixes the example to do what it says it does and use items.length as the stop value in the for loop.