Add clarification on frequencies (#26069)
After reading through the documentation it wasn't apparent (at least to me)
that the fundamental frequency was that of the oscillator's. By adding these
few lines we are explicitly stating that the fundamental frequency whose
associated component are specified through the `real` and `imag` arrays is
that of the oscillator (which, by default is 440 Hz).