[CSS building blocks] Clarify the location of content (#26071)
* Updated index.md

On line 114, it mentions forms being included in the HTML part of this course.  I'm in the middle of going through the content and I didn't encounter this so I thought I skipped over a section by mistake. I didn't. 

The content for the web forms section is further down from the CSS section.  I suggest that this paragraph should refer to the web forms section instead of the HTML section. This will be less confusing for new learners.

* Update files/en-us/learn/css/building_blocks/images_media_form_elements/index.md

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

---------

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>