`toISOString()` is wrong for `datetime-local` (#27579)
Update index.md

There aren't several methods to format a Date object for the datetime-local input. Either the user manually constructs a string in the correct format or they use a 3rd party library, and the example of using `toISOString()` is wrong. It will produce the wrong time in any timezone other than GMT.